Rename `language` module to `lang_prefs`
